+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 7.3.1
+## Added
+- support docxf and oform formats
+- create blank docxf from creation menu
+- create docxf from docx from creation menu
+- create oform from docxf from document manager
+
 ## 7.2.3
 ## Changed
 - opening editor in new tab for ownCloud Web